moving checkbox line component
