+++ inspector/src/components/actionTabs/tabs/propertyGrids/gui/imagePropertyGridComponent.tsx
@@ -6,7 +6,7 @@ import { LineContainerComponent } from "../../../lineContainerComponent";
 import { LockObject } from "../lockObject";
 import { Image } from "babylonjs-gui/2D/controls/image";
 import { FloatLineComponent } from "../../../lines/floatLineComponent";
-import { CheckBoxLineComponent } from "../../../lines/checkBoxLineComponent";
+import { CheckBoxLineComponent } from "../../../../../sharedUiComponents/lines/checkBoxLineComponent";
 import { OptionsLineComponent } from "../../../lines/optionsLineComponent";
 import { GlobalState } from '../../../../globalState';
