+++ inspector/src/components/actionTabs/tabs/propertyGrids/particleSystems/meshEmitterGridComponent.tsx
@@ -6,7 +6,7 @@ import { PropertyChangedEvent } from '../../../../propertyChangedEvent';
 import { LockObject } from '../lockObject';
 import { Vector3LineComponent } from '../../../lines/vector3LineComponent';
 import { MeshParticleEmitter } from 'babylonjs/Particles/EmitterTypes/meshParticleEmitter';
-import { CheckBoxLineComponent } from '../../../lines/checkBoxLineComponent';
+import { CheckBoxLineComponent } from '../../../../../sharedUiComponents/lines/checkBoxLineComponent';
 import { MeshPickerComponent } from '../../../lines/meshPickerComponent';
 import { Scene } from 'babylonjs/scene';
